Refactor Interception Vector plugin to remove activation logic and improve state handling; update controller button registration in WeaponSwitcher and TargetRecall plugins.

+++ NO_Tactitools/TargetRecall.cs
@@ -13,7 +13,7 @@ class TargetRecallPlugin {
             Plugin.inputCatcherPlugin.RegisterControllerButton(
                 Plugin.targetRecallControllerName.Value,
                 new ControllerButton(
-                (int)Plugin.targetRecallButtonNumber.Value,
+                Plugin.targetRecallButtonNumber.Value,
                 0.2f,
                 onShortPress: HandleClick,
                 onLongPress: HandleLongPress
